requirements.txt
matplotlib==3.5.0
numpy==1.19.2
Pillow==9.1.0
